GCP gcloud:: This resolved my 502 error response code.
I logged into one of my project instances on GCP.  Then I ran docker container inspect {containerid} to find that the -ENV ASPNETCORE_URLS=http://*:${PORT} ended in a ASPNETCORE_URLS="http://*:"  So no port.  I found in the GCP logs that the loadbalancer was trying to forward to 8080.  So I exposed 8080; not sure if that is needed because I think GCP passes that to docker.  But the ASPNETCORE_URLS does seem to be the issue.  So I hard coded it to 8080 and I was running.